requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyJWT==2.9.0
sqlmodel==0.0.22
httpx==0.27.2
openai==1.44.0
psycopg[binary]==3.2.9
email-validator==2.2.0
cryptography==43.0.3
alembic==1.13.2
google-auth==2.35.0
